component-entrypoints
- `avatar`: правки импортов (patch)
- `bottom-bar`: правки для `CleanGroup` (patch)
- `clean-buttons`: добавлена точка входа `index.ts` (patch)
- `clean-buttons`: CleanGroup теперь принимает произвольное содержимое (minor)
- `clean-buttons`: CleanGroupSizeContext заменен на CleanButtonContext (major)
